+++ lib/components/narrative/line-itin/line-itinerary.js
@@ -2,13 +2,13 @@ import coreUtils from '@opentripplanner/core-utils'
 import React from 'react'
 import styled from 'styled-components'
 
+import { ComponentContext } from '../../../util/contexts'
 import NarrativeItinerary from '../narrative-itinerary'
 import SaveTripButton from '../save-trip-button'
 import SimpleRealtimeAnnotation from '../simple-realtime-annotation'
-import { ComponentContext } from '../../../util/contexts'
 
-import ItinerarySummary from './itin-summary'
 import ItineraryBody from './connected-itinerary-body'
+import ItinerarySummary from './itin-summary'
 
 const { getLegModeLabel, getTimeZoneOffset, isTransit } = coreUtils.itinerary
 
@@ -19,16 +19,17 @@ export const LineItineraryContainer = styled.div`
 export default class LineItinerary extends NarrativeItinerary {
   static contextType = ComponentContext
 
-  _headerText () {
+  _headerText() {
     const { itinerary } = this.props
     return itinerary.summary || this._getSummary(itinerary)
   }
 
-  _getSummary (itinerary) {
+  _getSummary(itinerary) {
     let summary = ''
     const transitModes = []
     itinerary.legs.forEach((leg, index) => {
       if (isTransit(leg.mode)) {
+        /** TODO: remove use of deprecated unlocalized method once itinerary-body is localized */
         const modeStr = getLegModeLabel(leg)
         if (transitModes.indexOf(modeStr) === -1) transitModes.push(modeStr)
       }
@@ -36,6 +37,7 @@ export default class LineItinerary extends NarrativeItinerary {
 
     // check for access mode
     if (!isTransit(itinerary.legs[0].mode)) {
+      /** TODO: remove use of deprecated unlocalized method once itinerary-body is localized */
       summary += getLegModeLabel(itinerary.legs[0])
     }
 
@@ -47,7 +49,7 @@ export default class LineItinerary extends NarrativeItinerary {
     return summary
   }
 
-  render () {
+  render() {
     const {
       active,
       companies,
@@ -70,7 +72,7 @@ export default class LineItinerary extends NarrativeItinerary {
     }
 
     return (
-      <LineItineraryContainer className='line-itin'>
+      <LineItineraryContainer className="line-itin">
         <ItinerarySummary
           companies={companies}
           itinerary={itinerary}
@@ -81,8 +83,8 @@ export default class LineItinerary extends NarrativeItinerary {
         <SaveTripButton />
 
         {showRealtimeAnnotation && <SimpleRealtimeAnnotation />}
-        {active || expanded
-          ? <ItineraryBody
+        {active || expanded ? (
+          <ItineraryBody
             itinerary={itinerary}
             // Don't use setActiveLeg as an import
             // (will cause error when clicking on itinerary summary).
@@ -90,7 +92,7 @@ export default class LineItinerary extends NarrativeItinerary {
             setActiveLeg={setActiveLeg}
             timeOptions={timeOptions}
           />
-          : null}
+        ) : null}
         {ItineraryFooter && <ItineraryFooter />}
       </LineItineraryContainer>
     )
